requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
requests==2.32.4
python-dateutil==2.9.0
django-cors-headers==4.7.0